Migrate to new shared query response type
